[done] Human Intent Network. Add classes for: example, train, object
